requirements.txt
feedparser
requests
openai
python-dotenv
numpy